Update: Remove path awareness from data-views specific code.
