+++ packages/edit-site/src/components/app/index.js
@@ -17,7 +17,24 @@ import { GlobalStylesProvider } from '../global-styles/global-styles-provider';
 import DataviewsProvider from '../dataviews/provider';
 import { unlock } from '../../lock-unlock';
 
-const { RouterProvider } = unlock( routerPrivateApis );
+const { RouterProvider, useLocation } = unlock( routerPrivateApis );
+
+const PATH_TO_DATAVIEW_TYPE = {
+	'/pages': 'page',
+};
+
+function DataviewsProviderWithType( { children } ) {
+	const {
+		params: { path },
+	} = useLocation();
+	const viewType = PATH_TO_DATAVIEW_TYPE[ path ];
+	if ( ! viewType ) {
+		return <>{ children }</>;
+	}
+	return (
+		<DataviewsProvider type={ viewType }>{ children }</DataviewsProvider>
+	);
+}
 
 export default function App() {
 	const { createErrorNotice } = useDispatch( noticesStore );
@@ -39,10 +56,10 @@ export default function App() {
 			<GlobalStylesProvider>
 				<UnsavedChangesWarning />
 				<RouterProvider>
-					<DataviewsProvider>
+					<DataviewsProviderWithType>
 						<Layout />
 						<PluginArea onError={ onPluginAreaError } />
-					</DataviewsProvider>
+					</DataviewsProviderWithType>
 				</RouterProvider>
 			</GlobalStylesProvider>
 		</SlotFillProvider>
