+++ packages/edit-site/src/components/dataviews/provider.js
@@ -1,7 +1,6 @@
 /**
  * WordPress dependencies
  */
-import { privateApis as routerPrivateApis } from '@wordpress/router';
 import {
 	useEffect,
 	useState,
@@ -15,10 +14,7 @@ import { store as coreStore } from '@wordpress/core-data';
 /**
  * Internal dependencies
  */
-import { unlock } from '../../lock-unlock';
 import DataviewsContext from './context';
-
-const { useLocation } = unlock( routerPrivateApis );
 
 // DEFAULT_STATUSES is intentionally sorted. Items do not have spaces in between them.
 // The reason for that is to match the default statuses coming from the endpoint
@@ -44,10 +40,6 @@ const DEFAULT_VIEWS = {
 		hiddenFields: [ 'date', 'featured-image' ],
 		layout: {},
 	},
-};
-
-const PATH_TO_DATAVIEW_TYPE = {
-	'/pages': 'page',
 };
 
 function useDataviewTypeTaxonomyId( type ) {
@@ -183,7 +175,7 @@ function useDataviews( type, typeTaxonomyId ) {
 	return null;
 }
 
-function DataviewsProviderInner( { type, children } ) {
+function DataviewsProvider( { type, children } ) {
 	const [ currentViewId, setCurrentViewId ] = useState( null );
 	const dataviewTypeTaxonomyId = useDataviewTypeTaxonomyId( type );
 	const dataviews = useDataviews( type, dataviewTypeTaxonomyId );
@@ -241,21 +233,6 @@ function DataviewsProviderInner( { type, children } ) {
 			{ children }
 		</DataviewsContext.Provider>
 	);
-}
-function DataviewsProvider( { children } ) {
-	const {
-		params: { path },
-	} = useLocation();
-	const viewType = PATH_TO_DATAVIEW_TYPE[ path ];
-
-	if ( viewType ) {
-		return (
-			<DataviewsProviderInner type={ viewType }>
-				{ children }
-			</DataviewsProviderInner>
-		);
-	}
-	return <>{ children }</>;
 }
 
 let DataviewsProviderExported = Fragment;
